Split OS/RS test classes into separate modules.
